+++ src/Symfony/Component/HttpFoundation/Tests/Session/Storage/Handler/MemcachedSessionHandlerTest.php
@@ -119,7 +119,7 @@ class MemcachedSessionHandlerTest extends \PHPUnit_Framework_TestCase
 
     public function testGetConnection()
     {
-        $method = new \ReflectionMethod($this->storage, 'getConnection');
+        $method = new \ReflectionMethod($this->storage, 'getMemcached');
         $method->setAccessible(true);
  
         $this->assertInstanceOf('\Memcached', $method->invoke($this->storage));